docker_compose/empty_project/movies_admin/requirements.txt
uwsgi==2.0.22
django==3.2
flake8==4.0
python-dotenv==1.0
django-split-settings==1.2
psycopg2-binary==2.9.7
django-debug-toolbar==3.4.0